NOTEX

Inspect or change NOTEX.. NOTEX can be set for the entire table or for a specific variable.

Usage

px_notex(x, value)

# S3 method for class 'px'
px_notex(x, value)

Arguments

x

A px object

value

Optional. A character string, a data frame, or a list.

Use character, to set NOTEX for the entire table across all languages.

Use a data frame with columns 'language' and 'value' to set NOTEX for the entire table in a specific language.

Use a data frame with the columns 'variable-code' and 'notex', to set NOTEX for a specific variable across all languages. Add the column 'language' to set NOTEX for specific language.

Use a list of the above elements to set NOTEX in muliple ways. This is the same as calling NOTEX multiple times with different values.

If missing, the current NOTEX is returned.

If NULL, NOTEX is removed for the table and all variables.

Value

A px object, a character string, a data frame, or a list of character strings and/or data frames.