Save px object to file

Save px object to file

Usage

px_save(x, path, save_data = TRUE, data_path = NULL)

Arguments

x

A px object.

path

Path to file. The file extension determines the format. Can be:

.px to save as a px file

.xlsx to save as an Excel metadata workbook

save_data

If FALSE, no 'Data' sheet is created in the Excel workbook. Can only be used if path is an .xlsx file.

data_path

Path to an .rds file to save data table at. This is usefull when saving an Excel workbook where the data has more rows than Excel can handle. Can only be used if 'path' is an .xlsx file, and 'save_data' is TRUE.

Value

Nothing

Details

Use px_codepage() to change file encoding.

See also