Check all arguments to px_data

Check all arguments to px_data

Usage

validate_px_data_arguments(x, value, labels, validate)

Arguments

x

A px object

value

Optional. A data frame. If missing, the current DATA is returned. If NULL, all data rows are removed.

labels

Optional. Logic or character vector. If TRUE, the data table is returned with VALUES instead of CODES. By default the VALUES of the main language are returned, use a character language code to return VALUES for a specific language.

validate

Optional. If TRUE a number of validation checks are performed on the px object, and an error is thrown if the object is not valid. If FALSE, the checks are skipped, which can be usefull for large px objects where the check can be time consuming. Use px_validate() to manually preform the check.

Value

Nothing